Extract band-showing methods to new module, fix indentation
